requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
twilio==8.10.0
openai==1.3.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
sendgrid==6.10.0
gspread==5.12.0
oauth2client==4.1.3
pandas==2.1.3